requirements.txt
aiosmtpd==1.4.4.post2
arrow==1.3.0
asgiref==3.7.2
atpublic==4.0
attrs==23.2.0
binaryornot==0.4.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
cryptography==42.0.5
Django==4.1
django-browser-reload==1.12.1
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-tailwind==3.8.0
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
pillow==10.2.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-slugify==8.0.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
stripe==8.2.0
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.1.0
